+++ src/AuditTrail/Abstractions/IAuditTrailConsumer.cs
@@ -1,4 +1,5 @@
-﻿using AuditTrail.Models;
+﻿using AuditTrail.Enums;
+using AuditTrail.Models;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using System.Data.Common;
 
@@ -6,14 +7,11 @@ namespace AuditTrail.Abstractions;
 
 public interface IAuditTrailConsumer<TPermission>
 {
-    Task ConsumeAsync(IEnumerable<AuditTrailDataAfterSave<TPermission>> entities, CancellationToken cancellationToken = default);
+    Task ConsumeAsync(IEnumerable<AuditTrailDataAfterSave<TPermission>> auditTrailData, CancellationToken cancellationToken = default);
 
-    Task ConsumeTransactionAsync(IEnumerable<AuditTrailDataAfterSave<TPermission>> entities, TransactionEndEventData dbContextEventData, CancellationToken cancellationToken = default);
+    Task BeforeTransactionCommitedAsync(IEnumerable<AuditTrailDataAfterSave<TPermission>> auditTrailData, DbTransaction transaction, TransactionEventData eventData, CancellationToken cancellationToken = default);
 
-    virtual Task BeforeSaveAsync(IEnumerable<AuditTrailDataBeforeSave<TPermission>> entitiesCancellationToken, DbContextEventData eventData, CancellationToken cancellationToken = default)
-        => Task.CompletedTask;
-
-    virtual Task BeforeTransactionCommitedAsync(IEnumerable<AuditTrailDataAfterSave<TPermission>> auditTrailData, DbTransaction transaction, TransactionEventData eventData, CancellationToken cancellationToken = default)
+    Task TransactionFinished(TransactionEventData dbContextEventData, TransactionStatus status, CancellationToken cancellationToken = default)
         => Task.CompletedTask;
 }