+++ src/AuditTrail/Options/AuditTrailOptions.cs
@@ -1,0 +1,10 @@
+﻿namespace AuditTrail.Options;
+
+public class AuditTrailOptions
+{
+    /// <summary>
+    /// Autmatically open transaction for savechanges if no transaction is open
+    /// </summary>
+    public bool AutoOpenTransaction { get; set; }
+}
+