+++ src/AuditTrail/Services/AuditTrailService.cs
@@ -1,6 +1,8 @@
 ﻿using AuditTrail.Abstractions;
 using AuditTrail.Fluent.Abstractions;
+using AuditTrail.Options;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 
 namespace AuditTrail.Services;
 
@@ -8,8 +10,9 @@ public sealed class AuditTrailService<TPermission>(
     IAuditTrailConsumer<TPermission> AudtTrailConsumer,
     IServiceProvider ServiceProvider,
     IAuditTrailAssemblyProvider<TPermission> AuditAssemblyProvider,
-    ILogger<AuditTrailService<TPermission>> Logger) :
-    AuditTrailServiceBase<TPermission>(AudtTrailConsumer, ServiceProvider, AuditAssemblyProvider, Logger)
+    ILogger<AuditTrailService<TPermission>> Logger,
+    IOptions<AuditTrailOptions> options) :
+    AuditTrailServiceBase<TPermission>(AudtTrailConsumer, ServiceProvider, AuditAssemblyProvider, Logger, options)
 {
 }
 
@@ -17,8 +20,9 @@ public sealed class AuditTrailService<TPermission, TInstance>(
     IAuditTrailConsumer<TPermission, TInstance> AudtTrailConsumer,
     IServiceProvider ServiceProvider,
     IAuditTrailAssemblyProvider<TInstance> AuditAssemblyProvider,
-    ILogger<AuditTrailService<TPermission>> Logger) :
-    AuditTrailServiceBase<TPermission>(AudtTrailConsumer, ServiceProvider, AuditAssemblyProvider, Logger),
+    ILogger<AuditTrailService<TPermission>> Logger,
+    IOptions<AuditTrailOptions> options) :
+    AuditTrailServiceBase<TPermission>(AudtTrailConsumer, ServiceProvider, AuditAssemblyProvider, Logger, options),
     IAuditTrailService<TPermission, TInstance>
 {
     protected override object GetService(params Type[] types)